Moved spaces Legend modal to component, added to map
